+++ yamlparse-applicative/Changelog.md
@@ -2,14 +2,25 @@
 
 ## [Unreleased]
 
+### Added
+
+* `FieldParserOptional` and `FieldParserOptionalWithDefault`
+
+### Changed
+
+* `FieldParserOptional` is now called `FieldParserOptionalOrNull` and
+  `FieldParserOptionalWithDefault` is now called `FieldParserOptionalOrNullWithDefault`, but
+  if you don't use these constructors explicitly, no behaviour has changed.
+
+
 ## [0.1.0.2] - 2020-10-12
 
 ### Added
 
-+ `prettyColourisedSchemaDoc`
-+ `prettyColourisedParserDoc`
-+ `prettyColourisedSchema`
-+ `confDescWith`
+* `prettyColourisedSchemaDoc`
+* `prettyColourisedParserDoc`
+* `prettyColourisedSchema`
+* `confDescWith`
 
 
 ## [0.1.0.1] - 2020-05-10